+++ lib/middleman-aks/processor.rb
@@ -10,14 +10,44 @@ module Middleman
         @app = app
         @controller = controller
         @options = options
-        
+
+        bind_helpers
+        bind_after_configuration
+        bind_ready
       end
-      def logger
-        @app.logger
+      protected
+      def bind_helpers
+        if self.class.const_defined? :Helpers
+          klass = Object.const_get "#{self.class}::Helpers"
+          app.helpers do
+            $stderr.puts "* include helpers"
+            $stderr.puts klass.to_s
+            include klass
+          end
+        end
       end
-      # to be implemented in derived class
-      def manipulate_resource_list(resources)
-        resources
+      ## after_configuration hook
+      def bind_after_configuration  
+        processor = self
+        app.after_configuration do
+          processor.after_configuration if processor.respond_to? :after_configuration
+
+          if processor.respond_to? :manipulate_resource_list
+            name = processor.class.to_s.demodulize.underscore.to_sym
+            processor.app.sitemap.register_resource_list_manipulator(name, processor)
+          end
+        end
+      end
+      def bind_ready
+        ## ready hook
+        if respond_to? :ready
+          processor = self
+          app.ready do
+            $stderr.puts "* ready hook"
+            $stderr.puts processor.class.to_s
+            processor.ready
+          end
+        end
       end
     end ## class Processor
   end